// Client/front-end/src/components/TextArea/TextArea.tsx
import styled from "styled-components"

const TextArea = styled.textarea`
    background-color: #121212; 
   color: rgb(255, 255, 255);
   width: 100%;   
   margin: 8px 0;  
   padding: 12px 20px;   
   display: inline-block; 
   box-sizing: border-box;
   border-radius: 10px;
`
export default TextArea